refactoring: moved different keyword completers into `KeywordCompleter`
